+++ plutarch-ledger-api/CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
     * `PProposalProcedure`
     * `PGovernanceAction`
     * `PChangedParameters`
+    * `PProposalProcedure`
+* `PRationalData`, originally from `plutarch-extra`
 
 ### Modified